advanced/requirements.txt
# Core dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncio>=3.4.3

# AI/ML
openai>=1.0.0
anthropic>=0.25.0
tiktoken>=0.5.0

# Web scraping and search
beautifulsoup4>=4.12.0
playwright>=1.40.0
requests>=2.31.0
tavily-python>=0.3.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0

# Google integration
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
colorama>=0.4.6
rich>=13.0.0
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
ipython>=8.0.0